/tg/ Status Effects Prelude A - Updated Trait Backend (#4869)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Content from #4828 broken down in an individual PR

This does only a few things:
 * Adds /tg/ ASSERT which includes reasons
 * Updates the Traits backend from /tg/
* Refactors cloak cancellation with signals because it used
REMOVE_TRAIT_ALLSOURCES which is both bad and defunct

# Testing Photographs and Procedure
Tested trait backend as part of main PR