Fix imaginary friend ears and icons (#4863)
# About the pull request

This PR changes the imaginary friend ability to an orbit ability. The
reasoning for this is because if you have ghost ears set to only hear
nearby, merging would place you in the mob's contents so not in range to
hear anything.

This PR also fixes the hide ability not actually updating its icon
state.

# Explain why it's good for the game

Imaginary friends need to be able to hear who they are helping.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

https://youtu.be/95TtYocF2XQ

</details>


# Changelog
:cl: Drathek
fix: Fixed imaginary friend merge ability preventing hearing if ghost
ears are set to only nearby.
fix: Fixed imaginary friend hide ability not updating its icons.
/:cl: