Componentizes power cells (#4855)
# About the pull request
Creates a new, flexible component for power cells. Eventually, anything
powered will use a cell component instead of holding a ref to a
`/obj/item/cell` or having a `charge` variable.

Converts NVGs over to the new system as an example, the rest will come
in time.

Minor rework of how `emp_act()` works

# Explain why it's good for the game
Cells are one of the most obvious things to be componentized, and
componentizing them allows cells to be added to new objects in a very
easy manner.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Works

</details>