Should fix pumpkins, PUMPKINS EVERYWHERE (#4881)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Commit from
https://github.com/PvE-CMSS13/PvE-CMSS13/commit/0aa95b1d3692b63eed6b59ba5bca27db4f5faae6

Turns out i'm a monkey and didn't think about This One Small Thing when
refactoring, so the decorator keeps applying even without being active.
Neat huh.

# Explain why it's good for the game
Pumpkins overstepping their bounds. It's not much of a holiday if it's
on all the time...


# Changelog
:cl: MorrowWolf
fix: Fixed Pumpkins appearing past season.
/:cl:

Co-authored-by: Morrow <darthbane97@gmail.com>