Colonial Space Grunts (#4884)
# About the pull request

Adds the tabletop game known as Colonial Space Grunts, this will enable
roleplay for shipside roles when they have downtime.

yes i have permission to add this in game, it was done by a good friend
of mine who sadly doesnt play much anymore.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Colonial space grunts is a simple to learn and easy to get into tabletop
game, it can be played in a single game, ive ran several events with
this in the past and all of them have had positive reception, so i
figured id make it permenantly in game

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Adds Colonial Space Grunts to rec vend aswell as several pens and
Dice for playing the game.
/:cl: