Functional bedrolls (#4874)
# About the pull request

When playing on an event in a seperate server the event runner had
bedrolls we could lay out which made me wonder why we dont have them in
CM! So much like pill packets, with a bit of help here they are
functional bedrolls!

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Roleplay is good, i can see many situations where this could be used,
bedrolls around a campfire on sorokyne, making a more comfortable FOB,
pretending to go camping using the new tents, the opportunities are
endless
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>

https://imgur.com/a/O8m9gpc

# Changelog
:cl:
add: Adds functional bedrolls to the game.
/:cl: