Fix ability to fulton revivable bodies (#4933)
# About the pull request

Fixes the ability to fulton revivable bodies when the player inside the
body has ghosted.
Closes #4926

# Explain why it's good for the game

No more permaing in space or Normandy because an IO decided to kidnap
you via fulton


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
fix: Removes the ability to fulton bodies that can still be revived
/:cl: