LV-624 Cave Aesthetic Change (#4843)
# About the pull request

Alters the caves of LV-624 to have far more jungle grass and other
similar aesthetics. No layout changes are effected.

I hope to do more changes in the future, but I'll keep things separated
for now.

# Explain why it's good for the game

The caves of the map are quite bare in appearance, this is one small
chance to help increase the style of the map.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: LV-624's Caves have had an aesthetic change regarding jungle
tiles and flora.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>