Fixes some Almayer Walls randomly changing icons when hit (#4936)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Horizontal Almayer walls have "decoration" variants that are placed
randomly. Because these are applied in update_icon randomly, if you
shoot such a wall, it will cycle randomly through available icons.

On top of that, because the replacement icon was applied as an overlay,
it was drawn over everything else masking bullet and damage overlays.

This fixes both issues by making it use the icon instead and rolling
decorations ahead of time.

# Explain why it's good for the game
Consistency. Awfully distracting if you're in the firing range too

# Changelog
:cl:
fix: Horizontal Almayer walls no longer change icon randomly, and now
properly display damage overlays.
/:cl: