Moves sent faxes from highcom machines. (#4925)
# About the pull request
This PR makes it so that when using a fax machine that belongs to High
Command, the faxes appear in the correct category rather than in "Other"
category.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Makes it much easier to view faxes with consistency, and understand what
is going on for admins who may be replying.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Faxes sent from HighCom fax machines now appear in their correct
category.
/:cl: