Vehicles can be followed by ghosts (#4904)
# About the pull request

Loosens the type restrictions on what a ghost can orbit from `mob` to
`atom/movable`, allowing vehicles to be orbited again. Doesn't effect
ghosts viewing player UIs as that is still guarded by an `ishuman()`
check.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Vehicles are listed in a ghost's "Follow" menu but clicking them does
nothing. This PR allows the ghost to follow the vehicle, as expected.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/PvE-CMSS13/PvE-CMSS13/assets/14267245/70d96097-f3cb-4c99-bb98-f99bbf3de40d)


![image](https://github.com/PvE-CMSS13/PvE-CMSS13/assets/14267245/77e37a1c-6215-49f1-bdbd-61748b8ced59)

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: vehicles can be followed by ghosts
/:cl: