Adds Reflection to VV Modify Transform (#4907)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

As it says on the tin. Just gives the option to flip transform matrixes
in VV Modify Transform.

# Explain why it's good for the game
More shenanigans

# Testing Photographs and Procedure
Tested on brief chairs


# Changelog
:cl:
admin: VV 'Modify Transform' can now be used to flip objects.
/:cl: