Mess tech scaling (#4983)
# About the pull request

This PR adds scaling to allow for an extra mess tech slot once there are
70 marines playing.

This is a paid code request from a player.

# Explain why it's good for the game

Mess tech can be a lonely job and it can spice things up to have a
partner in the culinary arts.

# Testing Photographs and Procedure

I spawned in as a mess tech.

I then created 50 marines and checked if I could spawn in as a mess
tech. As expected the answer was no.

I then creates 20 more marines and checked if I could spawn in as a mess
tech and could.


# Changelog

:cl: Morrow
add: Mess tech positions now scale from 1 to 2 after 70 marines are in
the game
/:cl:

---------

Co-authored-by: BeagleGaming1 <56142455+BeagleGaming1@users.noreply.github.com>