Bureau du commandant - CO office (#4827)
# About the pull request

An update to the  CO office

# Explain why it's good for the game

If people like, its good, if people don't, it is not

# Testing Photographs and Procedure


![image](https://github.com/cmss13-devs/cmss13/assets/122557086/f4ca9cc9-9719-44bb-aef7-fcf95da5f7f5)


</details>


# Changelog
:cl:
maptweak: Changes to the CO office: The bathroom now faces into the CO's
bedroom, and not to his main office. Victory cabinet have been moved to
the Officer's mess because of logical issues. Extra-detail to the
office, including an exclusive stamp, table flags and others. The safe
is now inside the CO's bedroom, and not in his office. The energy APC is
now inside the CO's bedroom, and not in his office. Jones finally have a
BED again, or something like that. Other minor changes to objects, such
the addition of a cane, a box of glasses for serving guests, etc.
/:cl: