Break Up Secure Storage LV-624: V2 (#4944)
Breaks open the secure storage on LV-624. 

Remake of #4767 

Mapping changes are identical, however this PR has also removed buttons
that reference the now removed secure dome shutters from the rest of the
map, and similar nightmare inserts.

# Explain why it's good for the game

Survivor holds on 624 are usually a case of either the survivors running
and gunning or hiding in secure storage for 18 minutes until the xenos
reach T3, then its a case of if the Xenos can breach secure storage
before the Marines arrive. Secure storage is simply not a fun exercise
for both sides, if survivors are going to survive there should be more
effort than hiding in a pre-made bunker.

I should note that this does not make Secure Storage a impossible hold,
only that its now not 100% inaccessible to xenos until they reach T3.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: Secure Storage on LV-624 has been broken open, making it far
less defendable but easier to access.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>