Adds DISCORDURL to the example config (#5002)
# About the pull request

This adds a commented DISCORDURL entry to the example config in the
files, along with a comment explaining its use.

This entry is currently missing and is not mentioned anywhere in the
files so unless you already know what you are looking for, you are
forced to dig to get the exact string.

# Explain why it's good for the game

Config file clarity and general user friendliness is good.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


This does not seem like a front facing change at all, so did not include
a changelog.