Makes Thunderdome Floor Explosion-Proof (#4994)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Explosions could permanently mess with the tdome floor. This looks ugly
and happens frequently with the amount of abombs thrown in it.

# Explain why it's good for the game
Admin level looks cleaner and clearer.

# Testing Photographs and Procedure
Drop-Bomb enter enter

DMM edited in-place via regex, shouldn't be any mapmerging issues

# Changelog
:cl:
fix: The Thunderdome floor is now explosion-proof.
/:cl: