Fix resin doors not closing on mob/living if they are merged with weeds (#4990)
# About the pull request

This PR simply changes the mob/living check in
`/obj/structure/mineral_door/resin/Close()` to also account for whether
they are merged with the weeds. As soon as xeno corpses are also merging
with weeds this will handle them too.

# Explain why it's good for the game

Although I think this is a very rare situation, it can be especially
confusing if a fully weeded corpse that intentionally is hard to see is
causing a door to not close (seemingly arbitrarily).

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://github.com/cmss13-devs/cmss13/assets/76988376/4770ff65-c484-4e88-9381-19dfad22ffae

</details>


# Changelog
:cl: Drathek
fix: Resin doors will now close on dead mobs that are merged with weeds
(currently only human).
fix: Resin doors will now restart their closing timer each open making
the delay to close consistent.
code: Added a TRAIT_MERGED_WITH_WEEDS that is set whenever the mob is
currently merged with weeds.
/:cl: