Fix buried larva spawn grace period at start of round (#4988)
# About the pull request

Looking at the code again, #4502 is unintentionally eliminating the
buried larva grace period when there is no core. The first 30 minutes
should have a core be optional (because the hive is still being
established), but the queue requires a core (or nested bodies) to
dequeue players from the queue. So as soon as there is one player in
queue, buried spawns during the no-core grace period would be stopped.

This PR makes it so the queue will only prevent manual buried spawns
when there is a core.

# Explain why it's good for the game

Fixes a situation that needlessly hinders xeno start if the hive puts
off creating a core.

Also the logic was getting fairly complex here so the early returns
should make it more manageable.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
fix: Fixed buried larva spawn grace period at start of round if there is
a queue: Now join as xeno when there's a queue will only prevent buried
larva spawns if there is no core.
/:cl: