Lesser drones now die upon ghosting (#4978)
# About the pull request


https://forum.cm-ss13.com/t/ghosted-lesser-drones-should-just-die/4356/10

Lessers now die upon ghosting and are not offered to observers.

# Explain why it's good for the game

Nobody likes trying to join as xeno or see their position in the queue
only to see

Lesser Drone (420)
Lesser Drone (69)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
del: Lesser drones die upon ghosting and are not offered to observers.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>