Differentiates pylons in same area when spawning as lesser drone (#4985)
# About the pull request

Adds a (1) to duplicate named pylons if they reside in the same area
when trying to spawn at pylon as lesser drone.
Fixes #4424 

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Pylons now differentiated in the input list with a (1) if in the
same area.
/:cl: