Removal of Pizza Cutter from Req (#5023)
# About the pull request
I forgot how comically big this is with the in-hand sprite and it looks
really stupid so removing it. Can be re-added when someone makes a
proper sprite for it.

Partly revert of https://github.com/cmss13-devs/cmss13/pull/4998
# Explain why it's good for the game
I forgot how comically big this is with the in-hand sprite and it looks
really stupid so removing it. Can be re-added when someone makes a
proper sprite for it.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
del: Removed the Pizza Cutter.
/:cl: