Fixes Hijack Dropship launch announcement alarm not turning off - Bwoop is Kil (#5025)
# About the pull request

Turns the launch alarm off on dropships when the queen hijacks a ship.

# Explain why it's good for the game

Players found the alarm annoying when the ship was hijacked with the
alarm sounding and they couldn't turn it off.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty 

<details>
Happens in the first few seconds so you don't have to listen to youtube
and alarms and sensory overload


https://github.com/cmss13-devs/cmss13/assets/142365554/12485a52-a887-4ce0-825d-bb0b33e7e98d

</details>


# Changelog

:cl:
fix: Disables launch announcment alarm if it's sounding when queen
hijacks
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>