Kutjevo - South Caves Spring (#5039)
# About the pull request

Adds a small underground spring to the far south caves of Kutjevo and
slightly expands the existing cave system.

# Explain why it's good for the game

Kutjevo is a good map, but there's a lot of unused cave real estate
especially on the NE and SE corners of the map; so I thought this could
be a nice little addition which I plan to pair alongside other small
cave extensions in the future to hopefully just give it a bit more
personality and hopefully chip away at the unused space.

To help reduce the chances that it'll become some kind of meta hive spot
(somehow... even when it's right next to power plant...) I added a bunch
of large rocks ontop of the already proportionally large body of water
to reduce proper walling ability.

# Testing Photographs and Procedure

With Spring:

![image](https://github.com/cmss13-devs/cmss13/assets/31109792/59c3a316-39ac-4d52-8f18-823347eb2a6e)

Without Spring:

![image](https://github.com/cmss13-devs/cmss13/assets/31109792/d018d3b6-da45-499f-8980-3696f4143979)

Spring Close-Up:

![image](https://github.com/cmss13-devs/cmss13/assets/31109792/ed70d237-5b92-4ccd-8ca7-55c07864163b)

Short Video:

https://cdn.discordapp.com/attachments/746325498896056329/1178118032103587850/Kutjevo_Spring_Demo_720P.mp4?ex=6574faeb&is=656285eb&hm=9f1e047ab62ef4e99f9cabb682e6bd42599c1348c4cda0db4417de9a766b9ea7&
_(It will download if you click the link*)_

# Changelog
:cl:
mapadd: added a new spring area to kutjevo's south caves
/:cl: