added a file to handle the fridges in kitchen on almayer and added a bunch of food boxes. (#4586)
# About the pull request
i wanted to make it more clear code wise and separating the fridge used
on colony and other places from main kitchen.
the food boxes are there to help make stock version having less item
into them and make it cleaner.
made two version of each fridge on that hold less for the front kitchen
and one for the stock in the back room that hold way more.
remove a bunch of food crate that are duplicate with the vendor.
add a vendor that sell boxes of ingredients.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
better organization for kitchen and separation from other fridge allow
to not add item on top of fridge if you want to give him more...
cook will now be able to buy directly will hydroponic will stay viable
and they are still able to order or receive supply from req..

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added a lot's of food boxes for that hold the basic ingredients for
cooking.
add: Added a file to handle the fridges in almayer kitchen.
add: Added a vendor in the kitchen that sell boxes of ingredients.
add: Added a a food crate called surplus boxes ingredient containing
random boxes of ingredients.
add: Added a version of this crate that will be freely given via the ASR
system.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>