Removes flame particles (#5055)
# About the pull request

This is a paid PR.

Allegedly fire was lagging the player. I assume the particles are the
issue.

# Explain why it's good for the game

Given the vast range of people who play CM with varying levels of
computer hardware it would be prudent to remove anything that may be
lagging players. Game feel good.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
del: Removed flame particles. Possibly increases performance for some
players.
/:cl: