Eggsac Carrier Hardy weed fix. (#5053)
# About the pull request

In #4716 I forgot about hardy weeds. This allows carrier to plant the
fragile eggs on standard and hardy weeds which are basically standard
weeds +.

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Eggsac fragile eggs can be placed on hardy weeds.
/:cl: