Sanitizes burrower tunnel names to alphanumeric only (#4911)
# About the pull request

Someone made a tunnel with the characters `'🐝･ﾟ ･ﾟ·:｡･ﾟﾟ･` and the
tunnel was broken and not usable. This PR restricts them to alphanumeric
characters only.

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Restricts burrower tunnels to alphanumeric characters as some other
characters break the tunnel.
code; Adds new proc to replace non alphanumeric or space characters.
/:cl: