Consistent LZ names. (#5079)
# About the pull request

Alters the lz1/lz2 names of each map placed object to have LZ1 or LZ2 in
the front of the name, that way there's never any confusion over which
is which when looking at the DS consoles. "We're going to LZ2" should
never be a point of confusion for new POs anymore.

# Explain why it's good for the game

Information good.  


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: Standardized the names of LZs to include the name of the LZ.
/:cl: