Weeds are conserved under walls being upgraded and destroyed (#5096)
# About the pull request

Fixes #5060

Side effect of this PR is that when resin walls are destroyed then do
not remove the weeds underneath. However those weeds should still be
linked to a nearby node and recede as normal when that is destroyed.

# Explain why it's good for the game


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Upgraded resin walls can now nest hosts.
/🆑