Sorokyne Corporate Liaison (#4942)
# About the pull request

adds a corporate liaison survivor to sorokyne, along with a subtype of
the liaison's winter coat that can holster guns and a few other utility
items/weapons. it cannot hold any rifle magazines or similarly sized
items in its pockets though because there should be a tradeoff for how
cool it looks.

also fixed a very small typo in the political prisoner's name and
assignment.

# Explain why it's good for the game

more diversity in survivor types is good, map didn't have a corporate
survivor, a really cool sprite sees more use, fits with the map's lore,
etc.

the subtype of the liaison's winter coat is only available on this map
at the moment, and it's a downgrade from the survivor snowsuits/soviet
snowcoats that already spawn on sorokyne, so it's a fine tradeoff for it
looking cool i think. if needed i can always just make it a regular
liaison's winter coat.

typos are bad!!!

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/106038874/17c74652-5e36-495e-b48c-8a7c269b22df)

It's a cold day in Hell, suit.
</details>

# Changelog
:cl:IowaPotatoFarmer
add: Added a Corporate Liaison survivor to Sorokyne.
add: Added a modified version of the liaison's winter coat that allows
it to holster guns and a few other things. Only available to the
Sorokyne Strata Corporate Liaison for now.
spellcheck: Fixed a typo in the Sorokyne Strata Political Prisoner's ID
tag.
/:cl: