Eggsac carrier bugfix (#5059)
# About the pull request

So in one of various refactors of #4716 I removed some important code
from `Burst()` for the new egg type which removes it from the carrier's
sustained eggs. This is bad because we keep trying to call procs on eggs
which have already burst etc.

This removes the egg from the carrier's sustain list properly on
Burst().
I also added a check in Life() on the carrier to remove it from their
sustained eggs. If something funky happens we don't want to impact the
player from playing

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
🆑 
fix: Eggsack carrier eggs get properly removed from sustained list upon
bursting.
/🆑