Kutjevo Tunnels Upgrade (#5051)
# About the pull request

As part of my ongoing series of PRs to update tunnel placements,
especially on older maps, Kutjevo has had its original five tunnels
replaced with 7 new tunnels with more of a focus on the center and right
sides of the map.

# Explain why it's good for the game

Kutjevo's tunnels I find are mediocre at best, to start with there's a
very limited quantity of five tunnels, which was like Sorokyne until
recently, and of which three of them I believed to be mediocre at best
and the other two to be not particularly useful most times.

The scrubland tunnel is out in the open with no natural cover and it is
rather close to the LZ, this attempts to move it further backwards and
make it a little more out of the way.

The stony fields tunnel is not terrible but it suffers from a lack of
use since there was already a better option between botany and medical
and from personal experience I find the majority of landings are at the
north LZ rather than the south LZ. Since it's so out of the way it can
become a free shortcut to the comms tower in the late game should it
spawn at Hydroelectric.

The afforementioned botany-medical tunnel isn't bad, it's just plainly
in the open and it doesn't last particularly long once a foothold has
been established in medical or botany. The two existing tunnels inside
the caves are mediocre at best and from my personal experience any
normal hive tends to be just far enough away from them to make an
in-Hive tunnel a requirement.

I've opted to give both medical and botany their own "dedicated" tunnels
in the medical park and the NE corner of botany. I have moved the
scrubland tunnel further east behind some cover. The other tunnels
provide better access to the caves and the eastern construction area in
general.

# Testing Photographs and Procedure

(The lines are just visual aids to make finding the tunnel spots easier,
they are not indicative of other functions*)

Kutjevo's Existing Tunnels:
![Kutjevo Tunnel Map V2
0](https://github.com/cmss13-devs/cmss13/assets/31109792/addc07ed-c272-4e8c-94a1-2c25558b836b)

Kutjevo's New Tunnels:
![Kutjevo Tunnel Map V3
0](https://github.com/cmss13-devs/cmss13/assets/31109792/4fd97c14-4d62-4b64-961e-aaf61ba8b32e)

# Changelog

:cl:
mappadd: Added 7 new tunnels to Kutjevo
maptweak: Removed the 5 original tunnels on Kutjevo
/:cl: