Fixes a freeze when opening the 'Create Object' panel (#5125)
# About the pull request

Removes the long freeze when opening the 'Create Object' interface
through the game panel.
This was caused by line 40 of `create_object.html`, which attempts to
fill the object list with every subtype of `/obj` when the interface is
first opened. The "fix" was just removing that line.

https://github.com/cmss13-devs/cmss13/blob/0b133929724c569aff3945749c65559c4e65e4c5/html/create_object.html#L33-L42
This does mean that the user has to search for a type before anything
shows up, but I would consider that a pretty minor trade-off.
<hr>

I also increased the window size so that the 'Spawn' button is on the
screen by default, removed a couple of empty `value`s, and added some
spaces so that the HTML code lines up a bit more with the final UI.

# Explain why it's good for the game

Fixes a 30ish second freeze for admins. (Give or take a bit, but it's
still long either way.)

# Testing Photographs and Procedure
<details>
<summary>Videos</summary>

**Before:**


https://github.com/cmss13-devs/cmss13/assets/57483089/6d5d0aa7-74a9-4fbc-b78f-4bd6a960af87

**After:**


https://github.com/cmss13-devs/cmss13/assets/57483089/93fc0f99-41a5-4df6-9061-38740aa0e751

</details>


# Changelog
:cl:
admin: Fixed a freeze when opening the the 'Create Object' interface.
admin: Adjusted the size and positioning of the 'Create Object/Mob/Turf'
panels so that everything fits into the window.
/:cl: